Uploading version 1 to GitHub
This is to open source the project for anyone interested.  
This is code for an Arduino to accept text over serial and output it through motorshield driven solenoid valves as a scrolling bubble message.

Note the code is under CC BY-NC-SA licence which GitHub didn't offer.
This work is licensed under the Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License. To view a copy of this license, visit http://creativecommons.org/licenses/by-nc-sa/4.0/.